Move property editor files
