+++ src/Components/OATPropertyEditor/Internal/FormAddEnumItem.types.ts
@@ -1,5 +1,5 @@
 import { IDropdownOption } from '@fluentui/react';
-import { IOATPropertyEditorState } from './OATPropertyEditor.types';
+import { IOATPropertyEditorState } from '../OATPropertyEditor.types';
 
 export type ModalFormAddEnumItemProps = {
     languages: IDropdownOption[];
