+++ src/Components/OATPropertyEditor/Internal/FormUpdateProperty.types.ts
@@ -1,6 +1,6 @@
 import { IDropdownOption } from '@fluentui/react';
-import { IAction } from '../../Models/Constants/Interfaces';
-import { IOATPropertyEditorState } from './OATPropertyEditor.types';
+import { IAction } from '../../../Models/Constants/Interfaces';
+import { IOATPropertyEditorState } from '../OATPropertyEditor.types';
 
 export type FormUpdatePropertyProps = {
     dispatch: React.Dispatch<React.SetStateAction<IAction>>;
