+++ src/Components/OATPropertyEditor/Internal/JSONEditor.types.ts
@@ -1,0 +1,5 @@
+import { Theme } from '../../../Models/Constants/Enums';
+
+export type JSONEditorProps = {
+    theme?: Theme;
+};
