+++ src/Components/OATPropertyEditor/Internal/PropertiesModelSummary.types.ts
@@ -1,4 +1,4 @@
-import { IAction } from '../../Models/Constants/Interfaces';
+import { IAction } from '../../../Models/Constants/Interfaces';
 
 export type PropertiesModelSummaryProps = {
     dispatch: React.Dispatch<React.SetStateAction<IAction>>;
