+++ src/Components/OATPropertyEditor/Internal/PropertyListItem.tsx
@@ -1,30 +1,30 @@
 import React, { useState, useContext, useMemo } from 'react';
-import { CommandHistoryContext } from '../../Pages/OATEditorPage/Internal/Context/CommandHistoryContext';
+import { CommandHistoryContext } from '../../../Pages/OATEditorPage/Internal/Context/CommandHistoryContext';
 import { TextField, Text, IconButton } from '@fluentui/react';
 import {
     getPropertyEditorTextFieldStyles,
     getPropertyListItemIconWrapStyles,
     getPropertyListItemIconWrapMoreStyles,
     getPropertyInspectorStyles
-} from './OATPropertyEditor.styles';
-import { deepCopy } from '../../Models/Services/Utils';
+} from '../OATPropertyEditor.styles';
+import { deepCopy } from '../../../Models/Services/Utils';
 import PropertyListItemSubMenu from './PropertyListItemSubMenu';
 import { useTranslation } from 'react-i18next';
 import {
     SET_OAT_PROPERTY_EDITOR_CURRENT_PROPERTY_INDEX,
     SET_OAT_PROPERTY_MODAL_BODY,
     SET_OAT_PROPERTY_MODAL_OPEN
-} from '../../Models/Constants/ActionTypes';
+} from '../../../Models/Constants/ActionTypes';
 
 import {
     getModelPropertyCollectionName,
     getModelPropertyListItemName,
     getTargetFromSelection
-} from './Utils';
-import { FormBody } from './Constants';
+} from '../Utils';
+import { FormBody } from '../Shared/Constants';
 import { PropertyListItemProps } from './PropertyListItem.types';
-import { useOatPageContext } from '../../Models/Context/OatPageContext/OatPageContext';
-import { OatPageContextActionType } from '../../Models/Context/OatPageContext/OatPageContext.types';
+import { useOatPageContext } from '../../../Models/Context/OatPageContext/OatPageContext';
+import { OatPageContextActionType } from '../../../Models/Context/OatPageContext/OatPageContext.types';
 
 export const PropertyListItem: React.FC<PropertyListItemProps> = (props) => {
     const {
