+++ src/Components/OATPropertyEditor/Internal/PropertyListItem.types.ts
@@ -1,5 +1,5 @@
-import { DTDLProperty } from '../../Models/Classes/DTDL';
-import { IAction } from '../../Models/Constants/Interfaces';
+import { DTDLProperty } from '../../../Models/Classes/DTDL';
+import { IAction } from '../../../Models/Constants/Interfaces';
 
 export type PropertyListItemProps = {
     index?: number;
