+++ src/Components/OATPropertyEditor/Internal/PropertyListItemNest.types.ts
@@ -1,6 +1,6 @@
-import { DTDLProperty } from '../../Models/Classes/DTDL';
-import { IAction } from '../../Models/Constants';
-import { IOATLastPropertyFocused } from '../../Models/Constants/Interfaces';
+import { DTDLProperty } from '../../../Models/Classes/DTDL';
+import { IAction } from '../../../Models/Constants';
+import { IOATLastPropertyFocused } from '../../../Models/Constants/Interfaces';
 
 type IPropertySelectorTriggerElementsBoundingBox = {
     top: number;
