--- src/Components/OATPropertyEditor/JSONEditor.types.ts
@@ -1,5 +1,0 @@
-import { Theme } from '../../Models/Constants/Enums';
-
-export type JSONEditorProps = {
-    theme?: Theme;
-};
